feat(applications): migrate application details page to server/client architecture
